[MDH-158] refactor : ownerreservation 패키지 read, write 분리
