--- src/main/java/com/mdh/devtable/ownerreservation/presentation/dto/ShopReservationDateTimeSeatCreateRequest.java
@@ -1,4 +1,0 @@
-package com.mdh.devtable.ownerreservation.presentation.dto;
-
-public record ShopReservationDateTimeSeatCreateRequest() {
-}
