+++ src/main/java/com/mdh/devtable/ownerreservation/write/presentation/dto/SeatCreateRequest.java
@@ -1,4 +1,4 @@
-package com.mdh.devtable.ownerreservation.presentation.dto;
+package com.mdh.devtable.ownerreservation.write.presentation.dto;
 
 import com.mdh.devtable.reservation.domain.SeatType;
 import jakarta.validation.constraints.Max;
