+++ src/main/java/com/mdh/devtable/ownerreservation/write/presentation/dto/ShopReservationCreateRequest.java
@@ -1,4 +1,4 @@
-package com.mdh.devtable.ownerreservation.presentation.dto;
+package com.mdh.devtable.ownerreservation.write.presentation.dto;
 
 import com.mdh.devtable.reservation.domain.ShopReservation;
 import jakarta.validation.constraints.Max;
