+++ src/main/java/com/mdh/devtable/ownerreservation/write/presentation/dto/ShopReservationDateTimeSeatCreateRequest.java
@@ -1,0 +1,4 @@
+package com.mdh.devtable.ownerreservation.write.presentation.dto;
+
+public record ShopReservationDateTimeSeatCreateRequest() {
+}
